bugfix in OSC demo program: the OSC API expects a tempo value to be a float, not a double

+++ client/dev/osc/main.go
@@ -248,7 +248,7 @@ func loopPattern(pattern string) *osc.Bundle {
 	return bundle
 }
 
-func tempoMsg(offset int, bpm float64) *osc.Message {
+func tempoMsg(offset int, bpm float32) *osc.Message {
 	msg := osc.NewMessage("/system/tempo")
 	msg.Append(int32(offset))
 	msg.Append(bpm)
